requirements.txt
adamp==0.3.0
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
boto3==1.39.11
botocore==1.39.11
braceexpand==0.1.7
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
deepframe==0.1.1
deeplake==3.9.52
defusedxml==0.7.1
dill==0.4.0
eccv-caption==0.1.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humbug==0.3.2
idna==3.10
imageio==2.37.0
ipykernel==6.30.1
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.9
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
kiwisolver==1.4.9
lark==1.3.0
lazy-loader==0.4
libdeeplake==0.0.164
lz4==4.4.4
markupsafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mistune==3.1.4
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.7
notebook-shim==0.2.4
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathos==0.3.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.4.0
pox==0.3.6
ppft==1.7.7
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
send2trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.20
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
triton==3.4.0
types-python-dateutil==2.9.0.20250822
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==24.11.1
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.3
yarl==1.20.1
